updated logic for running forecasts, issue with api call based on input of longitude latitude etc

+++ src/main/java/com/EnergyForecasting/Model/ForecastOutput.java
@@ -10,7 +10,7 @@ public class ForecastOutput {
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(nullable = false, updatable = false)
     private Long outputID;
-    private Long interval;
+    private int interval;
     private String county;
     private Double offshoreOutput;
     private Double onshoreOutput;
@@ -20,7 +20,7 @@ public class ForecastOutput {
     @JoinColumn(name= "id", referencedColumnName = "id")
     private Forecast forecast;
 
-    public ForecastOutput(Long interval, String county, Double offshoreOutput, Double onshoreOutput, Double solarOutput, Forecast forecast) {
+    public ForecastOutput(int interval, String county, Double offshoreOutput, Double onshoreOutput, Double solarOutput, Forecast forecast) {
         this.interval = interval;
         this.county = county;
         this.offshoreOutput = offshoreOutput;
